Move tests outside of package
