--- satflow/tests/test_datasets.py
@@ -1,5 +1,0 @@
-from satflow.data.datasets import SatFlowDataset
-
-
-def test_dataset():
-    pass
